Enhance wiringPi integration and improve Python compatibility
- Updated CMakeLists.txt to include wiringPi headers from standard paths and source-build locations.
- Revised session documentation to clarify OS version and anticipated issues, including Python version compatibility.
- Fixed CLEDManager LED brightness logic to ensure correct active-low behavior.
- Adjusted main.cpp button mapping for diagnostics and updated sleep duration in update() to prevent I2C bus saturation.
- Modified KnightRider sequence timing for diagnostic purposes.
- Rewrote test_runner.py to improve compatibility with Python 3.5, including type hints and formatting changes.
- Added detailed troubleshooting notes for button tests and I2C communication.

+++ src/CLEDManager.cpp
@@ -76,9 +76,15 @@ void CLEDManager::ledBrightnessSet(int pLED, float pBrightness) {
 
         led->brightness = pBrightness;
 
-        // Convert power percentage to tick in the duty cycle
-        int powerScaledUp = pBrightness * 4095.0 / 100.0;
-
+        // Convert power percentage to tick in the duty cycle.
+        // LEDs are active-low: output HIGH = LED off, output LOW = LED on.
+        //   0%:   onTime=4095 → pwmWrite(4095) → HIGH 99.98% → LED off.               ✓
+        //   50%:  onTime=2048 → pwmWrite(2048) → HIGH 50%    → LED half on.            ✓
+        //   100%: onTime=0    → pwmWrite(0)    → pca9685FullOff (always LOW) → LED on. ✓
+        // WARNING: do NOT change onTime=0 → 4096 to "fix" brightness=100.
+        // pca9685FullOn (value >= 4096) = always HIGH = active-low LED OFF.
+        // pca9685FullOff (value == 0)   = always LOW  = active-low LED ON. This is correct.
+        int powerScaledUp = (int)(pBrightness * 4095.0 / 100.0);
         int onTime = 4095 - powerScaledUp;
 
         pwmWrite(led->pin, onTime);
